requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
openai==0.28
Werkzeug>=2.3.3
flask-cors
torch
transformers
Pillow  # For handling image inputs
safetensors
accelerate>=0.26.0
